Refactor DNS message handling and analytics pipeline
- Removed the old dnsmsg codec implementation and replaced it with a modular structure:
  - Introduced dnsmsg package with separate files for building responses (build.go), parsing queries and responses (parse.go), and modifying TTLs (ttl.go).

- Enhanced the circuit breaker logic in the forwarder to manage probe requests more effectively.

- Updated the forwarder to cache only A records, ensuring that AAAA records do not interfere with A record caching.

- Implemented a new analytics system to track query telemetry, ensuring it operates asynchronously to avoid blocking DNS responses.

- Streamlined the query execution pipeline in the rules package, improving error handling and logging for better observability.

- Improved RabbitMQ connection management with exponential backoff for reconnection attempts, enhancing resilience during broker outages.
