+++ DHCP/package.json
@@ -1,6 +1,6 @@
 {
   "name": "dhcp",
-  "version": "6.21.66-stable",
+  "version": "6.22.67-stable",
   "description": "this is the service to control all DHCP Related Works",
   "main": "DHCP.js",
   "scripts": {
