+++ VERSION
@@ -1,1 +1,1 @@
-6.21.66-stable
+6.22.67-stable
