+++ Web/README.md
@@ -44,14 +44,104 @@ shared/              infrastructure layer
 internal/            this service only
   config/            collection names and seed values
   app/               dependency wiring
-  dnsmsg/            DNS wire format encode/decode
+  dnsmsg/            DNS wire format
+    dnsmsg.go        Record type and shared constants
+    build.go         writing an answer packet
+    parse.go         reading a query or upstream response
+    ttl.go           rewriting TTLs in a formed response
   dnsio/             per-transport send/parse handlers
   netutil/           address discovery, SO_REUSEPORT sockets, IP scanner
   cache/ database/ dbpool/
   rules/             the query pipeline
+    rules.go         the StartRules type, wiring, cache invalidation
+    pipeline.go      the query path, layer by layer
+    analytics.go     per-query telemetry
+    servicestatus.go layer 1
+    blocklist.go     layer 2
   forwarder/         upstream forwarding, socket pool, circuit breakers
   server/            UDP, TCP and DoT listeners
 ```
+
+## Architecture
+
+The layering is ports-and-adapters. Every dependency arrow points **inward**:
+
+```
+   server/          →  rules/  →  dnsmsg/
+   (UDP/TCP/DoT)       (logic)    (wire format, depends on nothing)
+        ↓                 ↓
+     dnsio/          cache/ dbpool/ forwarder/
+   (the port)        (outbound adapters)
+
+              app/  ← composition root, wires it all once
+```
+
+Two consequences worth internalising:
+
+- **`dnsmsg` imports none of our packages.** It is pure byte manipulation, which
+  is why it is the one package with unit tests — it needs no running database.
+- **`rules` never mentions UDP, TCP or TLS.** It depends on the `dnsio.Handler`
+  interface (`internal/dnsio/handler.go`). That single indirection is why one
+  pipeline serves all three transports; adding a fourth means writing one new
+  adapter and touching nothing in `rules/`.
+
+## Follow one query end to end
+
+Start here when you need to find where something happens.
+
+| # | What happens | Where |
+|---|---|---|
+| 1 | Process starts, traps signals | `main.go:28` |
+| 2 | Every service constructed in dependency order | `internal/app/app.go:46` |
+| 3 | Infra connects, three transports bind | `internal/app/app.go:99` |
+| 4 | A datagram arrives; buffer is copied | `internal/server/udp.go:83` |
+| 5 | Query dispatched onto its own goroutine | `internal/server/udp.go:113` |
+| 6 | 5s timeout + panic guard wrap the query | `internal/rules/pipeline.go:33` |
+| 7 | The four layers run in order | `internal/rules/pipeline.go:65` |
+| 8 | **Layer 1** — DNS service on/off? | `internal/rules/servicestatus.go:80` |
+| 9 | **Layer 2** — is this domain blocked for this client? | `internal/rules/blocklist.go:63` |
+| 10 | **Layer 3** — local record: Redis, then Mongo | `internal/rules/pipeline.go:155` |
+| 11 | …Mongo lookup, following CNAMEs | `internal/dbpool/dbpool.go:51` |
+| 12 | **Layer 4** — no local answer, ask upstream | `internal/rules/pipeline.go:185` |
+| 13 | …shuffled resolver pool behind circuit breakers | `internal/forwarder/forwarder.go:97` |
+| 14 | Answer written back to the client socket | `internal/dnsio/udp.go:27` |
+| 15 | …the actual bytes of the reply packet | `internal/dnsmsg/build.go:12` |
+
+Steps 8–12 each either **answer and return**, or fall through to the next. That
+is the whole design; everything else is detail.
+
+## Go idioms used here
+
+If Go is not your daily language, these five account for most of what looks
+unfamiliar in this codebase.
+
+**`go someFunc()`** — runs `someFunc` concurrently and moves on immediately.
+`internal/server/udp.go:113` does this per datagram, so a slow query never blocks
+the socket read loop.
+
+**`defer`** — schedules a call for when the current function returns, no matter
+how it returns. Used for cleanup (`defer cancel()`) and for panic guards.
+
+**`recover()`** — catches a panic, Go's nearest equivalent to `catch`. It only
+works inside a `defer`, and only on its *own* goroutine — which is why
+`internal/rules/pipeline.go` installs one in `Execute` **and** another inside the
+goroutine it spawns. Miss the second and one bad packet kills the process.
+
+**`context.Context`** — the cancellation signal threaded through every call.
+`context.WithTimeout` gives the query 5 seconds; when it expires, every operation
+holding that context aborts. The `ctx` first parameter everywhere is this.
+
+**`chan` and `select`** — a channel is a typed pipe; `select` waits on several at
+once and takes whichever is ready first. `Execute` races the query against its
+timeout this way: whichever fires first wins.
+
+Two more that appear less often but matter:
+
+- **`atomic.Pointer[T]`** (`internal/dnsio/udp.go`) — lets the IP scanner swap
+  the socket out while queries are mid-flight, without a lock.
+- **`singleflight`** (`internal/rules/rules.go`) — if 500 queries for the same
+  domain miss the cache at once, exactly one database read happens and all 500
+  share its result.
 
 ## Query pipeline
 
@@ -105,14 +195,22 @@ bounded by memory, not by a limiter.
 
 ## Testing
 
-There are no tests yet. When adding them, note that Go requires `_test.go` files
-to live in the package directory they test. `internal/dnsmsg` is the place to
-start: its bounds checks replace behaviour JavaScript got from `try/catch`, so a
-mistake there means a malformed packet panics the server.
+There are no tests yet.
 
 ```bash
 go build ./... && go vet ./...
 ```
+
+When adding them, note that Go requires `_test.go` files to live in the package
+directory they test — they cannot go in the repo's `Test/` folder the way the
+Node suites do.
+
+`internal/dnsmsg` is the place to start: it depends on nothing (no database, no
+sockets), so it is testable in isolation, and its bounds checks replace behaviour
+JavaScript got from `try/catch` — a mistake there means a malformed packet panics
+the server. `internal/rules` is the natural second target, since its collaborators
+are already interfaces (`AnalyticsPublisher`, `dnsio.Handler`) and can be faked
+without standing up Mongo or Redis.
 
 ## Verifying a running server
