+++ server/package.json
@@ -1,6 +1,6 @@
 {
   "name": "nexoraldns-server",
-  "version": "6.21.66-stable",
+  "version": "6.22.67-stable",
   "description": "this server is responsible for managing the all records that are custom build for the all networks",
   "keywords": [
     "server"
